Update handing of missing #permitted_attributes.
- Merge Actions::ResourceMethods into Actions::ResourceAction.
