+++ lib/cuprum/rails/rspec/actions/update_contracts.rb
@@ -112,13 +112,8 @@ module Cuprum::Rails::RSpec::Actions
           # :nocov:
         end
 
-        include_contract 'resource action contract'
-
-        include_contract(
-          'should require permitted attributes',
-          params: configured_params,
-          &should_not_update_the_entity
-        )
+        include_contract 'resource action contract',
+          require_permitted_attributes: true
 
         include_contract(
           'should require primary key',
